resolved comments from PR

+++ client/components/checkout.js
@@ -32,11 +32,7 @@ class DisconnectedCheckout extends React.Component {
     })
 
     axios.post('/api/orders', {orderItems: orders})
-    if(this.props.user) {
-        axios.delete('/api/cart', {data: {userId: this.props.user.id}})
-    } else {
-        
-    }
+    if(this.props.user) axios.delete('/api/cart', {data: {userId: this.props.user.id}})
     this.setState({cartItems: []})
   }
